+++ commu-intell/ros/setup.bash
@@ -5,9 +5,5 @@ chmod +x src/beginner_tutorials/srv/*
 chmod +x src/beginner_tutorials/scripts/*
 chmod +x src/imu/src/scripts/*
 chmod +x src/motor/src/scripts/*
-<<<<<<< HEAD
 chmod +x src/gps_pose/src/*
 chmod +x src/gps_pose/scripts/*
-=======
-chmod +x src/gps_raw_data/src/*
->>>>>>> b3b87ceede747d2224a05706eb002b5974f11b70
