+++ commu-intell/ros/src/gps_pose/CMakeLists.txt
@@ -12,6 +12,7 @@ find_package(catkin REQUIRED COMPONENTS
   rospy
   std_msgs
   message_generation
+  message_runtime
 )
 
 ## System dependencies are found with CMake's conventions
@@ -56,7 +57,7 @@ find_package(catkin REQUIRED COMPONENTS
 ## Generate services in the 'srv' folder
  add_service_files(
    FILES
-   GpsPose.srv
+   GetGpsPose.srv
  )
 
 ## Generate actions in the 'action' folder
@@ -70,6 +71,7 @@ find_package(catkin REQUIRED COMPONENTS
  generate_messages(
    DEPENDENCIES
    std_msgs
+ #  message_generation
  )
 
 ################################################
