+++ commu-intell/ros/src/gps_pose/src/gps_raw_data_client.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 import sys
 import rospy
 from gps_pose.srv import *
@@ -5,7 +7,7 @@ from gps_pose.srv import *
 def gps_raw_data_client():
     rospy.wait_for_service('gps_raw_data')
     try:
-        gps_raw_data = rospy.ServiceProxy('gps_raw_data', GpsPoseResponse)
+        gps_raw_data = rospy.ServiceProxy('gps_raw_data', GetGpsPose)
         resp = gps_raw_data()
         return resp.gps
     except rospy.ServiceException, e:
@@ -22,4 +24,4 @@ if __name__ == "__main__":
     gps_data = gps_raw_data_client()
     print('Latitude: {}'.format(gps_data.latitude))
     print('Longitude: {}'.format(gps_data.longitude))
-    print('Altitute: {}'.format(gps_data.altitute))
+    print('Altitute: {}'.format(gps_data.altitude))
