+++ commu-intell/ros/src/gps_pose/srv/GetGpsPose.srv
@@ -1,0 +1,2 @@
+---
+GpsPose gps
